+++ hoot-core/src/main/cpp/hoot/core/visitors/BuildingHeightVisitor.h
@@ -22,16 +22,16 @@
  * This will properly maintain the copyright information. Maxar
  * copyrights will be updated automatically.
  *
- * @copyright Copyright (C) 2019, 2020, 2021 Maxar (http://www.maxar.com/)
+ * @copyright Copyright (C) 2019, 2020, 2021, 2022 Maxar (http://www.maxar.com/)
  */
 
 #ifndef BUILDING_HEIGHT_VISITOR_H
 #define BUILDING_HEIGHT_VISITOR_H
 
 // hoot
-#include <hoot/core/visitors/ConstElementVisitor.h>
-#include <hoot/core/info/NumericStatistic.h>
 #include <hoot/core/criterion/BuildingCriterion.h>
+#include <hoot/core/info/NumericStatistic.h>
+#include <hoot/core/visitors/ConstElementVisitor.h>
 
 // Qt
 #include <QRegularExpression>
@@ -66,7 +66,7 @@ public:
   double getMax() const override { return _maxHeight; }
   double getAverage() const override
   {
-    const double average = _numAffected == 0 ? 0.0 : _totalHeight / _numAffected;
+    const double average = _numAffected == 0 ? 0.0 : _totalHeight / static_cast<double>(_numAffected);
     return average;
   }
