+++ hoot-js/src/main/cpp/hoot/js/elements/ElementJs.cpp
@@ -166,7 +166,7 @@ void ElementJs::getId(const FunctionCallbackInfo<Value>& args)
 
   ConstElementPtr e = ObjectWrap::Unwrap<ElementJs>(args.This())->getConstElement();
 
-  args.GetReturnValue().Set(Integer::New(current, e->getId()));
+  args.GetReturnValue().Set(Integer::New(current, static_cast<int32_t>(e->getId())));
 }
 
 void ElementJs::getStatusInput(const FunctionCallbackInfo<Value>& args)
